requirements.txt
aiohttp==3.9.0 ; python_full_version == "3.11.5"
aiosignal==1.3.1 ; python_full_version == "3.11.5"
annotated-types==0.6.0 ; python_full_version == "3.11.5"
anyio==3.7.1 ; python_full_version == "3.11.5"
attrs==23.1.0 ; python_full_version == "3.11.5"
beautifulsoup4==4.12.2 ; python_full_version == "3.11.5"
certifi==2023.11.17 ; python_full_version == "3.11.5"
charset-normalizer==3.3.2 ; python_full_version == "3.11.5"
colorama==0.4.6 ; python_full_version == "3.11.5" and platform_system == "Windows"
dataclasses-json==0.6.2 ; python_full_version == "3.11.5"
diskcache==5.6.3 ; python_full_version == "3.11.5"
flaml==2.1.1 ; python_full_version == "3.11.5"
frozenlist==1.4.0 ; python_full_version == "3.11.5"
greenlet==3.0.1 ; (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_full_version == "3.11.5"
idna==3.4 ; python_full_version == "3.11.5"
jsonpatch==1.33 ; python_full_version == "3.11.5"
jsonpointer==2.4 ; python_full_version == "3.11.5"
langchain==0.0.337 ; python_full_version == "3.11.5"
langsmith==0.0.65 ; python_full_version == "3.11.5"
marshmallow==3.20.1 ; python_full_version == "3.11.5"
multidict==6.0.4 ; python_full_version == "3.11.5"
mypy-extensions==1.0.0 ; python_full_version == "3.11.5"
numpy==1.26.2 ; python_full_version == "3.11.5"
openai==0.28.1 ; python_full_version == "3.11.5"
packaging==23.2 ; python_full_version == "3.11.5"
pyautogen==0.1.14 ; python_full_version == "3.11.5"
pydantic-core==2.14.3 ; python_full_version == "3.11.5"
pydantic==2.5.1 ; python_full_version == "3.11.5"
python-dotenv==1.0.0 ; python_full_version == "3.11.5"
pyyaml==6.0.1 ; python_full_version == "3.11.5"
requests==2.31.0 ; python_full_version == "3.11.5"
sniffio==1.3.0 ; python_full_version == "3.11.5"
soupsieve==2.5 ; python_full_version == "3.11.5"
sqlalchemy==2.0.23 ; python_full_version == "3.11.5"
tenacity==8.2.3 ; python_full_version == "3.11.5"
termcolor==2.3.0 ; python_full_version == "3.11.5"
tqdm==4.66.1 ; python_full_version == "3.11.5"
typing-extensions==4.8.0 ; python_full_version == "3.11.5"
typing-inspect==0.9.0 ; python_full_version == "3.11.5"
urllib3==2.1.0 ; python_full_version == "3.11.5"
yarl==1.9.2 ; python_full_version == "3.11.5"
